+++ test/testXliffSplit.js
@@ -150,6 +150,43 @@ module.exports.xliffsplit = {
         test.equal(actual, expected);
         test.done();
     },
+    testXliffSplitdistritueSerialize3_xliffStyle: function(test) {
+        test.expect(2);
+        var settings = {};
+        settings.xliffVersion = 2;
+        settings.infiles = [
+            "testfiles/xliff20/merge-en-US-style.xliff",
+        ];
+        settings.xliffStyle = "custom"
+        var superset = XliffSplit(settings);
+        var result = XliffSplit.distribute(superset, settings);
+        test.ok(result);
+
+        var actual = result["app2"].serialize();
+        var expected =
+        '<?xml version="1.0" encoding="utf-8"?>\n' +
+        '<xliff version="2.0" srcLang="en-KR" trgLang="en-US" xmlns:l="http://ilib-js.com/loctool">\n' +
+        '  <file id="app2_f1" original="app2">\n' +
+        '    <group id="app2_g1" name="javascript">\n' +
+        '      <unit id="app2_1">\n' +
+        '        <segment>\n' +
+        '          <source>app2: String 2a</source>\n' +
+        '          <target>app2: String 2a</target>\n' +
+        '        </segment>\n' +
+        '      </unit>\n' +
+        '      <unit id="app2_2">\n' +
+        '        <segment>\n' +
+        '          <source>app2: String 2b</source>\n' +
+        '          <target>app2: String 2b</target>\n' +
+        '        </segment>\n' +
+        '      </unit>\n' +
+        '    </group>\n' +
+        '  </file>\n' +
+        '</xliff>';
+
+        test.equal(actual, expected);
+        test.done();
+    },
     testXliffSplitWrite: function(test) {
         test.expect(3);
         rmrf("testfiles/xliff20/splitTest/app1/en-US.xliff");